requirements_streamlit.txt
streamlit==1.37.0
requests==2.32.3
trafilatura==1.12.2
readability-lxml==0.8.1
beautifulsoup4==4.12.3
lxml==5.3.0
pydub==0.25.1
openai>=1.30.0
python-dotenv==1.0.1
plotly==5.17.0
streamlit-option-menu==0.3.6
streamlit-lottie==0.0.5
pyaudioop>=0.4.2
